refactor(viewer): extract Cesium asset plugin + harden dev middleware
Addresses PR #1179 review:
- Codex: vite.config.ts crossed the AGENTS.md ~400-line split limit. Move the
  Cesium plugin into apps/viewer/vite-plugins/cesium-assets.ts (config drops to
  328 lines). The plugin now resolves the output dir from config.root instead of
  __dirname, so it's independent of its own file location.
- CodeRabbit: the dev /cesium middleware's containment check used a bare string
  prefix (a sibling like …/Cesium-backup could pass). Compare on a path-separator
  boundary instead.
- CodeRabbit: guard the existsSync→createReadStream TOCTOU window with an 'error'
  handler so a vanished/locked file can't crash the dev server.

No behaviour change to the production copy (still recursive fs.cp of the four
subdirs flat under <outDir>/cesium); dev-only hardening otherwise.
